+++ src/main/java/duke/task/Event.java
@@ -1,13 +1,17 @@
+package duke.task;
+
 import java.time.LocalDateTime;
+import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
-import java.time.format.FormatStyle;
 
-public class Deadline extends Task {
+public class Event extends Task {
     private final LocalDateTime dateAndTime;
+    private final LocalTime endTime;
 
-    public Deadline(String taskName, LocalDateTime dateAndTime, boolean isDone) {
+    public Event(String taskName, LocalDateTime dateAndTime, LocalTime endTime, boolean isDone) {
         super(taskName, isDone);
         this.dateAndTime = dateAndTime;
+        this.endTime = endTime;
     }
 
     /**
@@ -16,12 +20,13 @@ public class Deadline extends Task {
      */
     @Override
     public String toString() {
-        String dateAndTime = DateTimeFormatter.ofPattern("dd MMM yyyy 'at' h:mm a")
+        String dateAndTime = DateTimeFormatter.ofPattern("dd MMM yyyy 'from' h:mm a")
                 .format(this.dateAndTime);
+        String endTime = DateTimeFormatter.ofPattern("h:mm a").format(this.endTime);
         if (super.isDone) {
-            return "[D][X] " + super.taskName + " (by: " + dateAndTime + ")";
+            return "[E][X] " + super.taskName + " (at: " + dateAndTime + " - " + endTime + ")";
         } else {
-            return "[D][ ] " + super.taskName + " (by: " + dateAndTime + ")";
+            return "[E][ ] " + super.taskName + " (at: " + dateAndTime + " - " + endTime + ")";
         }
     }
 
@@ -32,7 +37,8 @@ public class Deadline extends Task {
     @Override
     public String parseToString() {
         String result = super.isDone ? "X" : "0";
-        result += "D/" + super.taskName + "/" + this.dateAndTime.toString();
+        result += "E/" + super.taskName + "/" + this.dateAndTime.toString() + "/"
+                + this.endTime.toString();
         return result;
     }
 }
