+++ src/main/java/duke/task/Task.java
@@ -1,3 +1,5 @@
+package duke.task;
+
 public abstract class Task {
     protected String taskName;
     protected boolean isDone;
